CS237B_Win25_HW1-main/requirements.txt
numpy
torch==2.3.0
torchvision
torchaudio
matplotlib
tqdm
Pillow